Revisi Sprint 2 : - Mengubah dan menampilakan Created by dan Modified By di Master Specialization - Menyesuaikan tombol Chat dokter dan kondisi offline dengan hasil pencarian dokter - memperbaiki database pada m_doctor_medical_facility_category

+++ HealthCare340B.API/Controllers/SpecializationDoctorController.cs
@@ -15,7 +15,7 @@ namespace HealthCare340B.API.Controllers
             dASpecializationDoctor = new DASpecializationDoctor(_db);
         }
         [HttpGet("{id}")]
-        public async Task<ActionResult> GetById(int id)
+        public async Task<ActionResult> GetById(long id)
         {
             try
             {
